move some tests to payment_methods_test.go
