+++ pkg/godax/godax_test.go
@@ -149,125 +149,6 @@ func TestClient_ListFills(t *testing.T) {
 	}
 }
 
-// FIXME: payment_methods_test.go
-func TestClient_ListPaymentMethods(t *testing.T) {
-	tests := [...]struct {
-		name    string
-		fields  fields
-		want    []PaymentMethod
-		wantRaw string
-		wantErr bool
-	}{
-		{
-			name:   "when a successful call is made to list payment methods",
-			fields: defaultFields(),
-			want: []PaymentMethod{{
-				ID:            "bc6d7162-d984-5ffa-963c-a493b1c1370b",
-				Type:          "ach_bank_account",
-				Name:          "Bank of America - eBan... ********7134",
-				Currency:      "USD",
-				PrimaryBuy:    true,
-				PrimarySell:   true,
-				AllowBuy:      true,
-				AllowSell:     true,
-				AllowDeposit:  true,
-				AllowWithdraw: true,
-				Limits: PMLimits{
-					Buy: []PMLimit{{
-						PeriodInDays: 1,
-						Total: PMAmount{
-							Amount:   "10000.00",
-							Currency: "USD",
-						},
-						Remaining: PMAmount{
-							Amount:   "10000.00",
-							Currency: "USD",
-						},
-					}},
-					Sell: []PMLimit{{
-						PeriodInDays: 1,
-						Total: PMAmount{
-							Amount:   "10000.00",
-							Currency: "USD",
-						},
-						Remaining: PMAmount{
-							Amount:   "10000.00",
-							Currency: "USD",
-						},
-					}},
-				},
-			}},
-			wantErr: false,
-			wantRaw: `[{
-				"id": "bc6d7162-d984-5ffa-963c-a493b1c1370b",
-				"type": "ach_bank_account",
-				"name": "Bank of America - eBan... ********7134",
-				"currency": "USD",
-				"primary_buy": true,
-				"primary_sell": true,
-				"allow_buy": true,
-				"allow_sell": true,
-				"allow_deposit": true,
-				"allow_withdraw": true,
-				"limits": {
-					"buy": [{
-						"period_in_days": 1,
-						"total": {
-							"amount": "10000.00",
-							"currency": "USD"
-						},
-						"remaining": {
-							"amount": "10000.00",
-							"currency": "USD"
-						}
-					}],
-					"sell": [{
-						"period_in_days": 1,
-						"total": {
-							"amount": "10000.00",
-							"currency": "USD"
-						},
-						"remaining": {
-							"amount": "10000.00",
-							"currency": "USD"
-						}
-					}]
-				}
-			}]`,
-		},
-	}
-	for _, tt := range tests {
-		mockClient := MockResponse(tt.wantRaw)
-
-		t.Run(tt.name, func(t *testing.T) {
-			c := &Client{
-				baseRestURL: tt.fields.baseRestURL,
-				baseWsURL:   tt.fields.baseWsURL,
-				key:         tt.fields.key,
-				secret:      tt.fields.secret,
-				passphrase:  tt.fields.passphrase,
-				httpClient:  mockClient,
-			}
-
-			got, err := c.ListPaymentMethods()
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Client.ListPaymentMethods() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if len(c.httpClient.(*MockClient).Requests) != 1 {
-				t.Errorf("should have made one request, but made: %d", len(c.httpClient.(*MockClient).Requests))
-			}
-
-			validateHeaders(t, c)
-
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("Client.ListPaymentMethods() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
 // TODO: accounts_test.go ???
 func TestClient_ListCoinbaseAccounts(t *testing.T) {
 	tests := [...]struct {
@@ -395,62 +276,6 @@ func TestClient_ListCoinbaseAccounts(t *testing.T) {
 	}
 }
 
-// FIXME: fees_test.go
-func TestClient_GetCurrentFees(t *testing.T) {
-	tests := [...]struct {
-		name    string
-		fields  fields
-		want    Fees
-		wantRaw string
-		wantErr bool
-	}{
-		{
-			name:   "when a successful call is made to get current fees",
-			fields: defaultFields(),
-			want: Fees{
-				MakerFeeRate: "0.0015",
-				TakerFeeRate: "0.0025",
-				USDVolume:    "25000.00",
-			},
-			wantRaw: `{
-				"maker_fee_rate": "0.0015",
-				"taker_fee_rate": "0.0025",
-				"usd_volume": "25000.00"
-			}`,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			mockClient := MockResponse(tt.wantRaw)
-
-			c := &Client{
-				baseRestURL: tt.fields.baseRestURL,
-				baseWsURL:   tt.fields.baseWsURL,
-				key:         tt.fields.key,
-				secret:      tt.fields.secret,
-				passphrase:  tt.fields.passphrase,
-				httpClient:  mockClient,
-			}
-
-			got, err := c.GetCurrentFees()
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Client.GetCurrentFees() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if len(c.httpClient.(*MockClient).Requests) != 1 {
-				t.Errorf("should have made one request, but made: %d", len(c.httpClient.(*MockClient).Requests))
-			}
-
-			validateHeaders(t, c)
-
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("Client.GetCurrentFees() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
 // TODO: figure file
 func TestClient_GetTrailingVolume(t *testing.T) {
 	tests := [...]struct {
@@ -505,188 +330,6 @@ func TestClient_GetTrailingVolume(t *testing.T) {
 			if !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("Client.GetTrailingVolume() = %v, want %v", got, tt.want)
 			}
-		})
-	}
-}
-
-// FIXME: profiles_test.go
-func TestClient_ListProfiles(t *testing.T) {
-	tests := [...]struct {
-		name    string
-		fields  fields
-		want    []Profile
-		wantRaw string
-		wantErr bool
-	}{
-		{
-			name:   "when a successful call is made to list profiles",
-			fields: defaultFields(),
-			want: []Profile{{
-				ID:        "86602c68-306a-4500-ac73-4ce56a91d83c",
-				UserID:    "5844eceecf7e803e259d0365",
-				Name:      "default",
-				Active:    true,
-				IsDefault: true,
-				CreatedAt: "2019-11-18T15:08:40.236309Z",
-			}},
-			wantRaw: `[{
-				"id": "86602c68-306a-4500-ac73-4ce56a91d83c",
-				"user_id": "5844eceecf7e803e259d0365",
-				"name": "default",
-				"active": true,
-				"is_default": true,
-				"created_at": "2019-11-18T15:08:40.236309Z"
-			}]`,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			mockClient := MockResponse(tt.wantRaw)
-
-			c := &Client{
-				baseRestURL: tt.fields.baseRestURL,
-				baseWsURL:   tt.fields.baseWsURL,
-				key:         tt.fields.key,
-				secret:      tt.fields.secret,
-				passphrase:  tt.fields.passphrase,
-				httpClient:  mockClient,
-			}
-
-			got, err := c.ListProfiles()
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Client.ListProfiles() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if len(c.httpClient.(*MockClient).Requests) != 1 {
-				t.Errorf("should have made one request, but made: %d", len(c.httpClient.(*MockClient).Requests))
-			}
-
-			validateHeaders(t, c)
-
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("Client.ListProfiles() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-// FIXME: profiles_test.go
-func TestClient_GetProfile(t *testing.T) {
-	type args struct {
-		profileID string
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		args    args
-		want    Profile
-		wantRaw string
-		wantErr bool
-	}{
-		{
-			name:   "when a successful call is made to get a profile by ID",
-			fields: defaultFields(),
-			args:   args{profileID: "123abc"},
-			want: Profile{
-				ID:        "86602c68-306a-4500-ac73-4ce56a91d83c",
-				UserID:    "5844eceecf7e803e259d0365",
-				Name:      "default",
-				Active:    true,
-				IsDefault: true,
-				CreatedAt: "2019-11-18T15:08:40.236309Z",
-			},
-			wantRaw: `{
-				"id": "86602c68-306a-4500-ac73-4ce56a91d83c",
-				"user_id": "5844eceecf7e803e259d0365",
-				"name": "default",
-				"active": true,
-				"is_default": true,
-				"created_at": "2019-11-18T15:08:40.236309Z"
-			}`,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			mockClient := MockResponse(tt.wantRaw)
-
-			c := &Client{
-				baseRestURL: tt.fields.baseRestURL,
-				baseWsURL:   tt.fields.baseWsURL,
-				key:         tt.fields.key,
-				secret:      tt.fields.secret,
-				passphrase:  tt.fields.passphrase,
-				httpClient:  mockClient,
-			}
-
-			got, err := c.GetProfile(tt.args.profileID)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Client.GetProfile() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if len(c.httpClient.(*MockClient).Requests) != 1 {
-				t.Errorf("should have made one request, but made: %d", len(c.httpClient.(*MockClient).Requests))
-			}
-
-			validateHeaders(t, c)
-
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("Client.GetProfile() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-// TODO: profiles_test.go
-func TestClient_ProfileTransfer(t *testing.T) {
-	type args struct {
-		transfer TransferParams
-	}
-	tests := [...]struct {
-		name    string
-		fields  fields
-		args    args
-		wantErr bool
-		wantRaw string
-	}{
-		{
-			name:   "when a successful market order is made to transfer currency across profiles",
-			fields: defaultFields(),
-			args: args{transfer: TransferParams{
-				From:     "86602c68-306a-4500-ac73-4ce56a91d83c",
-				To:       "e87429d3-f0a7-4f28-8dff-8dd93d383de1",
-				Currency: "BTC-USD",
-				Amount:   "0.05",
-			}},
-			wantErr: false,
-			wantRaw: "",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			mockClient := MockResponse(tt.wantRaw)
-
-			c := &Client{
-				baseRestURL: tt.fields.baseRestURL,
-				baseWsURL:   tt.fields.baseWsURL,
-				key:         tt.fields.key,
-				secret:      tt.fields.secret,
-				passphrase:  tt.fields.passphrase,
-				httpClient:  mockClient,
-			}
-
-			err := c.ProfileTransfer(tt.args.transfer)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Client.ProfileTransfer() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if len(c.httpClient.(*MockClient).Requests) != 1 {
-				t.Errorf("should have made one request, but made: %d", len(c.httpClient.(*MockClient).Requests))
-			}
-
-			validateHeaders(t, c)
 		})
 	}
 }
@@ -1181,70 +824,6 @@ func TestClient_Get24HourStatsForProduct(t *testing.T) {
 
 			if !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("Client.Get24HourStatsForProduct() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-// FIXME: currencies_test.go
-func TestClient_ListCurrencies(t *testing.T) {
-	tests := []struct {
-		name    string
-		fields  fields
-		want    []Currency
-		wantRaw string
-		wantErr bool
-	}{
-		{
-			name:   "when a successful call is made to list currencies",
-			fields: defaultFields(),
-			want: []Currency{{
-				ID:      "BTC",
-				Name:    "Bitcoin",
-				MinSize: "0.00000001",
-			}, {
-				ID:      "USD",
-				Name:    "United States Dollar",
-				MinSize: "0.01000000",
-			}},
-			wantRaw: `[{
-				"id": "BTC",
-				"name": "Bitcoin",
-				"min_size": "0.00000001"
-			}, {
-				"id": "USD",
-				"name": "United States Dollar",
-				"min_size": "0.01000000"
-			}]`,
-		},
-	}
-	for _, tt := range tests {
-		mockClient := MockResponse(tt.wantRaw)
-
-		t.Run(tt.name, func(t *testing.T) {
-			c := &Client{
-				baseRestURL: tt.fields.baseRestURL,
-				baseWsURL:   tt.fields.baseWsURL,
-				key:         tt.fields.key,
-				secret:      tt.fields.secret,
-				passphrase:  tt.fields.passphrase,
-				httpClient:  mockClient,
-			}
-
-			got, err := c.ListCurrencies()
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Client.ListCurrencies() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if len(c.httpClient.(*MockClient).Requests) != 1 {
-				t.Errorf("should have made one request, but made: %d", len(c.httpClient.(*MockClient).Requests))
-			}
-
-			validateHeaders(t, c)
-
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("Client.ListCurrencies() = %v, want %v", got, tt.want)
 			}
 		})
 	}
